move SolverSubprocessError to errors.py
